Moving report form to homepage
