+++ src/components/form-report/form-report.scss
@@ -1,0 +1,58 @@
+form-report {
+  #form-report {
+    #form-step-2 {
+      h2 {
+        margin-top: 0.5rem;
+      }
+
+      #file-input-button {
+        float: left;
+      }
+
+      .photo-preview-container {
+        display: inline-block;
+        position: relative;
+
+        .delete {
+          position: absolute;
+          top: -0.5rem;
+          right: -0.5rem;
+          z-index: 2;
+          cursor: pointer;
+          background-color: rgba(0, 0, 0, 0.5);
+
+          &:hover {
+            background-color: rgba(0, 0, 0, 0.75);
+          }
+        }
+
+        #photo-preview {
+          opacity: 0.85;
+          margin-left: 1.5rem;
+          width: 40px;
+          height: 40px;
+          border-radius: $radius;
+          background-size: cover;
+          background-position: center;
+          object-fit: cover;
+        }
+      }
+    }
+  }
+
+  .agreement {
+    font-size: 13px;
+
+    stencil-route-link a {
+      color: $blue;
+      text-decoration: underline;
+    }
+  }
+
+  .social-radio-group,
+  .report-watchdog-distributor-radio-group {
+    .radio {
+      margin-bottom: 0;
+    }
+  }
+}
